# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpusZero API",
    "description": "Enterprise-grade API for the OpusZero event management and experience booking platform",
    "version": "1.0.0",
    "contact": {
      "name": "OpusZero API Support",
      "email": "api@opuszero.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://opuszero.com/license"
    }
  },
  "servers": [
    {
      "url": "https://api.opuszero.com",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.opuszero.com",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/events": {
      "get": {
        "summary": "List events",
        "description": "Retrieve a list of events with optional filtering",
        "tags": ["Events"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["draft", "published", "cancelled"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create event",
        "description": "Create a new event",
        "tags": ["Events"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "description": "Response data"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["success"]
      },
...
      "CreateEventRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Summer Music Festival"
          },
          "slug": {
            "type": "string",
            "example": "summer-music-festival"
          },
          "description": {
            "type": "string"
          },
          "start_date": {
            "type": "string",
            "format": "date-time"
          },
          "end_date": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "published"],
            "default": "draft"
          }
        },
        "required": ["name", "slug", "status"]
      },
...
      "EventsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
      "Event": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start_date": {
            "type": "string",
            "format": "date-time"
          },
          "end_date": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "published", "cancelled"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}